git-svn-id: http://svn.seqan.de/seqan/trunk@4627 e6417c60-b987-48fd-844e-b20f0fcc1017

+++ seqan/projects/library/seqan/graph_msa/graph_align_tcoffee_library.h
@@ -421,8 +421,8 @@ __getAlignmentStatistics(String<TFragment, TSpec1>& matches,
 	getAlignmentStatistics(matches, pairSet, (TPos) from, (TPos) length(matches),  matchLen, overlapLen, alignLen);
 			
 	// Calculate sequence similarity
-	if (i < j) dist[i*nseq+j] = SEQAN_DISTANCE_UNITY - (TValue) (((double) matchLen / (double) overlapLen) * ((double) overlapLen / (double) alignLen) * SEQAN_DISTANCE_UNITY);
-	else dist[j*nseq+i] = SEQAN_DISTANCE_UNITY - (TValue) (((double) matchLen / (double) overlapLen) * ((double) overlapLen / (double) alignLen) * SEQAN_DISTANCE_UNITY);
+	if (i < j) dist[i*nseq+j] = SEQAN_DISTANCE_UNITY - (TValue) (((double) matchLen / (double) overlapLen) * ((double) overlapLen / (double) alignLen) * (double) SEQAN_DISTANCE_UNITY);
+	else dist[j*nseq+i] = SEQAN_DISTANCE_UNITY - (TValue) (((double) matchLen / (double) overlapLen) * ((double) overlapLen / (double) alignLen) * (double) SEQAN_DISTANCE_UNITY);
 }
 
 //////////////////////////////////////////////////////////////////////////////
